code verfier avec sonar

+++ src/main/java/tn/esprit/spring/services/EmployeServiceImpl.java
@@ -3,6 +3,7 @@ package tn.esprit.spring.services;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -37,42 +38,64 @@ public class EmployeServiceImpl implements IEmployeService {
 	}
 
 	public void mettreAjourEmailByEmployeId(String email, int employeId) {
-		Employe employe = employeRepository.findById(employeId).get();
-		employe.setEmail(email);
-		employeRepository.save(employe);
+		Optional<Employe> optionalemploye = employeRepository.findById(employeId);
+		if (optionalemploye.isPresent())
+		{
+			Employe employe = optionalemploye.get();
+			employe.setEmail(email);
+			employeRepository.save(employe);
+		}
+		
 
 	}
-
+	
 	@Transactional	
 	public void affecterEmployeADepartement(int employeId, int depId) {
-		Departement depManagedEntity = deptRepoistory.findById(depId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
+Optional<Departement> optionldepartement = deptRepoistory.findById(depId);
+		
+		Optional<Employe> optionalemploye = employeRepository.findById(employeId);
+		
 
-		if(depManagedEntity.getEmployes() == null){
-
+		if(optionldepartement.isPresent() &&  optionalemploye.isPresent() )
+		{
+			Departement depManagedEntity = optionldepartement.get();
+			Employe employeManagedEntity = optionalemploye.get();
+		if (depManagedEntity.getEmployes() == null)
+		{
 			List<Employe> employes = new ArrayList<>();
 			employes.add(employeManagedEntity);
 			depManagedEntity.setEmployes(employes);
-		}else{
+		}
+		else{
 
 			depManagedEntity.getEmployes().add(employeManagedEntity);
 
 		}
-
+			
+		}
 	}
 	@Transactional
 	public void desaffecterEmployeDuDepartement(int employeId, int depId)
 	{
-		Departement dep = deptRepoistory.findById(depId).get();
-
-		int employeNb = dep.getEmployes().size();
-		for(int index = 0; index < employeNb; index++){
-			if(dep.getEmployes().get(index).getId() == employeId){
-				dep.getEmployes().remove(index);
-				break;//a revoir
+		
+Optional<Departement> optionaldepartement =deptRepoistory.findById(depId);
+		
+		
+		if (optionaldepartement.isPresent())
+		{
+			Departement dep = optionaldepartement.get();
+			
+			int employeNb = dep.getEmployes().size();
+			for(int index = 0; index < employeNb; index++){
+				if(dep.getEmployes().get(index).getId() == employeId){
+					dep.getEmployes().remove(index);
+					break;
+				}
 			}
 		}
+		
 	}
+
 
 	public int ajouterContrat(Contrat contrat) {
 		contratRepoistory.save(contrat);
@@ -80,37 +103,51 @@ public class EmployeServiceImpl implements IEmployeService {
 	}
 
 	public void affecterContratAEmploye(int contratId, int employeId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
-
-		contratManagedEntity.setEmploye(employeManagedEntity);
-		contratRepoistory.save(contratManagedEntity);
+Optional<Contrat> optionalcontrat = contratRepoistory.findById(contratId);
 		
-	}
-
-	public String getEmployePrenomById(int employeId) {
-		Employe employeManagedEntity = employeRepository.findById(employeId).get();
-		return employeManagedEntity.getPrenom();
-	}
-	public void deleteEmployeById(int employeId)
-	{
-		Employe employe = employeRepository.findById(employeId).get();
-
-		//Desaffecter l'employe de tous les departements
-		//c'est le bout master qui permet de mettre a jour
-		//la table d'association
-		for(Departement dep : employe.getDepartements()){
-			dep.getEmployes().remove(employe);
+		Optional<Employe> optionalemploye = employeRepository.findById(employeId);
+		
+		if (optionalcontrat.isPresent() && optionalemploye.isPresent())
+		{
+			Contrat contratManagedEntity = optionalcontrat.get();
+			Employe employeManagedEntity = optionalemploye.get();
+			contratManagedEntity.setEmploye(employeManagedEntity);
+			contratRepoistory.save(contratManagedEntity);
 		}
 
-		employeRepository.delete(employe);
+		
+		
 	}
-
-	public void deleteContratById(int contratId) {
-		Contrat contratManagedEntity = contratRepoistory.findById(contratId).get();
-		contratRepoistory.delete(contratManagedEntity);
-
+	public String getEmployePrenomById(int employeId) {
+Optional<Employe> optionalemploye = employeRepository.findById(employeId);
+		
+		if (optionalemploye.isPresent())
+		{
+			Employe employeManagedEntity = optionalemploye.get();
+			return employeManagedEntity.getPrenom();
+		}
+		else return "employe prénom n'existe pas";
 	}
+	
+	public void deleteEmployeById(int employeId)
+	{
+		Optional<Employe> optionalemploye = employeRepository.findById(employeId);
+		if (optionalemploye.isPresent())
+		{
+			Employe employe = optionalemploye.get();
+
+		
+			for(Departement dep : employe.getDepartements()){
+				dep.getEmployes().remove(employe);
+			}
+
+			employeRepository.delete(employe);
+		}
+	}
+		
+
+
+		
 
 	public int getNombreEmployeJPQL() {
 		return employeRepository.countemp();
